None fix env variables yml (#215)
* fix action percentage reading from yml file

+++ app/util/post_run/cleanup_results_dir.py
@@ -7,7 +7,8 @@ FILES_TO_REMOVE = ['jmeter.out',
                    'merged.json',
                    'merged.yml',
                    'PyTestExecutor.ldjson',
-                   'system.properties']
+                   'system.properties',
+                   'locust.out']
 
 if ENV_TAURUS_ARTIFACT_DIR in os.environ:
     artifacts_dir = os.environ.get(ENV_TAURUS_ARTIFACT_DIR)
@@ -16,9 +17,10 @@ else:
 
 for file in FILES_TO_REMOVE:
     file_path = Path(f'{artifacts_dir}/{file}')
-    try:
-        os.remove(file_path)
-        print(f'The {file} was removed successfully')
-    except OSError as e:
-        print(f'Deleting of the {file} failed!\n'
-              f'Error: {file_path}: {e.strerror}')
+    if file_path.exists():
+        try:
+            os.remove(file_path)
+            print(f'The {file} was removed successfully')
+        except OSError as e:
+            print(f'Deleting of the {file} failed!\n'
+                  f'Error: {file_path}: {e.strerror}')
